refactor: 입력한 숫자를 검증하고 ApplauseGame에 필요한 paramter로 바꾸어 저장하는 ApplauseParameter를 구현, 테스트코드 작성. 적절히 값 메서드 이동. 테스트 통과.
